requirements.txt
Django==2.2.18
psycopg2==2.8.4
djangorestframework==3.12.2
